Added Art Assets
Hey guys I changed the background up a bit, and it is still basically a
placeholder, but I think once we work out the size and scope of the map
I will be able to make things like the sand and landscape and everything
make sense. I plan on designing it in photoshop and all in one file, and
then we can design the buildings and such to overlay on top of it.